@@ -85,55 +87,55 @@
     "griddle-overhaul-subgrid-plugin": "*",
     "html-loader": "^0.4.3",
     "http-proxy-rules": "^1.0.1",
-    "immutable": "^3.7.6",
+    "immutable": "^3.8.1",
     "imports-loader": "^0.6.5",
-    "jquery": "^2.2.3",
+    "jquery": "^2.2.4",
     "jwt-simple": "^0.5.0",
-    "less": "^2.6.1",
+    "less": "^2.7.1",
     "less-loader": "^2.2.3",
     "memory-fs": "^0.3.0",
     "ncp": "^2.0.0",
     "nodemon": "^1.9.2",
-    "npm-run-all": "^2.1.0",
-    "pmx": "^0.6.1",
-    "postcss-browser-reporter": "^0.4.0",
+    "npm-run-all": "^2.1.1",
+    "pmx": "^0.6.2",
+    "postcss-browser-reporter": "^0.5.0",
     "postcss-cssnext": "^2.5.2",
-    "postcss-loader": "^0.8.2",
+    "postcss-loader": "^0.9.1",
     "postcss-reporter": "^1.3.3",
-    "postcss-url": "^5.1.1",
+    "postcss-url": "^5.1.2",
     "raw-loader": "^0.5.1",
-    "react": "^15.0.2",
-    "react-bootstrap": "^0.29.2",
-    "react-cookie": "^0.4.5",
+    "react": "^15.1.0",
+    "react-bootstrap": "^0.29.4",
+    "react-cookie": "^0.4.6",
     "react-dom": "^15.0.2",
     "react-hatch": "*",
-    "react-redux": "^4.4.2",
-    "react-redux-idle-monitor": "^0.1.17",
-    "react-router": "^2.4.0",
+    "react-redux": "^4.4.5",
+    "react-redux-idle-monitor": "^0.3.1",
+    "react-router": "^2.4.1",
     "react-router-redux": "^4.0.4",
-    "react-tooltip": "^1.2.1",
+    "react-tooltip": "^2.0.2",
     "react-transform-catch-errors": "^1.0.2",
     "react-transform-hmr": "^1.0.4",
     "react-transform-render-visualizer": "^0.4.0",
-    "redbox-react": "^1.2.3",
+    "redbox-react": "^1.2.5",
     "redux": "^3.5.2",
     "redux-actions": "^0.9.1",
     "redux-addons": "*",
     "redux-blueprint": "*",
-    "redux-devtools": "^3.2.0",
+    "redux-devtools": "^3.3.1",
     "redux-devtools-chart-monitor": "^1.4.1",
     "redux-devtools-themes": "^1.0.0",
-    "redux-form": "^5.2.3",
+    "redux-form": "^5.2.4",
     "redux-grid": "*",
     "redux-grid-view": "*",
-    "redux-idle-monitor": "^0.3.1",
+    "redux-idle-monitor": "^0.3.2",
     "redux-middleware": "^0.1.20",
     "redux-mux": "*",
-    "regenerator": "^0.8.42",
-    "repackage": "^0.2.15",
+    "regenerator": "^0.8.46",
+    "repackage": "^0.2.17",
     "router": "^1.1.4",
     "safe-json-stringify": "^1.0.3",
-    "script-loader": "^0.6.1",
+    "script-loader": "^0.7.0",
     "serialize-javascript": "^1.2.0",
     "serve-favicon": "^2.3.0",
     "serve-file": "^0.1.4",
@@ -143,68 +145,68 @@
   },
   "devDependencies": {
     "autoprefixer": "6.3.6",
-    "babel-cli": "^6.7.0",
-    "babel-core": "^6.7.0",
-    "babel-eslint": "^6.0.2",
+    "babel-cli": "^6.9.0",
+    "babel-core": "^6.9.0",
+    "babel-eslint": "^6.0.4",
     "babel-loader": "^6.2.4",
     "babel-plugin-react-transform": "^2.0.2",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-polyfill": "^6.7.4",
-    "babel-preset-es2015": "^6.6.0",
+    "babel-polyfill": "^6.9.0",
+    "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.5.0",
     "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-stage-0": "^6.5.0",
     "bootstrap-loader": "^1.0.10",
     "bootstrap-sass": "^3.3.6",
-    "classnames": "^2.2.3",
+    "classnames": "^2.2.5",
     "css-loader": "^0.23.1",
-    "esdoc": "^0.4.6",
+    "esdoc": "^0.4.7",
     "esdoc-es7-plugin": "0.0.3",
-    "eslint": "^2.7.0",
-    "eslint-config-airbnb": "6.2.0",
-    "eslint-plugin-react": "^4.3.0",
-    "expect": "^1.16.0",
+    "eslint": "^2.10.2",
+    "eslint-config-airbnb": "9.0.1",
+    "eslint-plugin-react": "^5.1.1",
+    "expect": "^1.20.1",
     "exports-loader": "^0.6.3",
     "expose-loader": "^0.7.1",
     "express": "^4.13.4",
     "express-open-in-editor": "^1.2.0",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.8.5",
-    "history": "^2.0.1",
+    "history": "^2.1.1",
     "imports-loader": "^0.6.5",
     "jasmine-core": "^2.4.1",
     "json-loader": "^0.5.4",
     "karma": "^0.13.22",
-    "karma-chrome-launcher": "^0.2.3",
-    "karma-mocha": "^0.2.2",
+    "karma-chrome-launcher": "^1.0.1",
+    "karma-mocha": "^1.0.1",
     "karma-webpack": "^1.7.0",
-    "less": "^2.6.1",
+    "less": "^2.7.1",
     "less-loader": "^2.2.3",
-    "mocha": "^2.4.5",
+    "mocha": "^2.5.1",
     "morgan": "^1.7.0",
-    "node-sass": "^3.4.2",
-    "postcss-import": "^8.1.0",
-    "postcss-loader": "^0.8.2",
-    "react-addons-css-transition-group": "^15.0.1",
+    "node-sass": "^3.7.0",
+    "postcss-import": "^8.1.2",
+    "postcss-loader": "^0.9.1",
+    "react-addons-css-transition-group": "^15.1.0",
     "react-document-meta": "^2.0.3",
-    "react-dom": "^15.0.1",
+    "react-dom": "^15.1.0",
     "react-hot-loader": "^1.3.0",
     "react-loading-order-with-animation": "^1.0.0",
     "react-transform-hmr": "^1.0.4",
     "redux-devtools-dock-monitor": "^1.1.1",
-    "redux-form": "^5.0.1",
+    "redux-form": "^5.2.4",
     "redux-logger": "2.6.1",
-    "redux-thunk": "^2.0.1",
+    "redux-thunk": "^2.1.0",
     "resolve-url-loader": "^1.4.3",
     "rimraf": "^2.5.2",
     "sass-loader": "^3.2.0",
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
-    "webpack": "^1.12.14",
+    "webpack": "^1.13.1",
     "webpack-dev-middleware": "^1.6.1",
     "webpack-dev-server": "^1.14.1",
     "webpack-hot-middleware": "^2.10.0",
-    "webpack-merge": "^0.9.0"
+    "webpack-merge": "^0.12.0"
   },
   "author": "Cole Chamberlain <cole.chamberlain@gmail.com> (https://github.com/cchamberlain)",
   "bugs": {
