@@ -1,12 +1,12 @@
 webpackJsonp([7,1],{
 
-/***/ 418:
+/***/ 421:
 /*!**************************************************************!*\
   !*** ../src/app/routes/routes/NoMatch/components/NoMatch.js ***!
   \**************************************************************/
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _redboxReact2 = __webpack_require__(/*! redbox-react */ 31);\n\nvar _redboxReact3 = _interopRequireDefault(_redboxReact2);\n\nvar _reactTransformCatchErrors3 = __webpack_require__(/*! react-transform-catch-errors */ 28);\n\nvar _reactTransformCatchErrors4 = _interopRequireDefault(_reactTransformCatchErrors3);\n\nvar _react2 = __webpack_require__(/*! react */ 1);\n\nvar _react3 = _interopRequireDefault(_react2);\n\nvar _reactTransformHmr3 = __webpack_require__(/*! react-transform-hmr */ 29);\n\nvar _reactTransformHmr4 = _interopRequireDefault(_reactTransformHmr3);\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar _components = {\n  NoMatch: {\n    displayName: 'NoMatch'\n  }\n};\n\nvar _reactTransformHmr2 = (0, _reactTransformHmr4.default)({\n  filename: 'C:/Users/ColeChamberlain/cchamberlain/redux-webpack-boilerplate/src/app/routes/routes/NoMatch/components/NoMatch.js',\n  components: _components,\n  locals: [module],\n  imports: [_react3.default]\n});\n\nvar _reactTransformCatchErrors2 = (0, _reactTransformCatchErrors4.default)({\n  filename: 'C:/Users/ColeChamberlain/cchamberlain/redux-webpack-boilerplate/src/app/routes/routes/NoMatch/components/NoMatch.js',\n  components: _components,\n  locals: [],\n  imports: [_react3.default, _redboxReact3.default]\n});\n\nfunction _wrapComponent(id) {\n  return function (Component) {\n    return _reactTransformHmr2(_reactTransformCatchErrors2(Component, id), id);\n  };\n}\n\n/**\r\n * NoMatch Module\r\n * @module NoMatch\r\n */\n\nvar NoMatch = _wrapComponent('NoMatch')(function (_Component) {\n  _inherits(NoMatch, _Component);\n\n  function NoMatch() {\n    _classCallCheck(this, NoMatch);\n\n    return _possibleConstructorReturn(this, Object.getPrototypeOf(NoMatch).apply(this, arguments));\n  }\n\n  _createClass(NoMatch, [{\n    key: 'render',\n    value: function render() {\n      return _react3.default.createElement('div', null);\n    }\n  }]);\n\n  return NoMatch;\n}(_react2.Component));\n\nexports.default = NoMatch;\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(/*! ./../../../../../../~/webpack/buildin/module.js */ 26)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ../src/app/routes/routes/NoMatch/components/NoMatch.js\n ** module id = 418\n ** module chunks = 7\n **/\n//# sourceURL=webpack:///../src/app/routes/routes/NoMatch/components/NoMatch.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _redboxReact2 = __webpack_require__(/*! redbox-react */ 32);\n\nvar _redboxReact3 = _interopRequireDefault(_redboxReact2);\n\nvar _reactTransformCatchErrors3 = __webpack_require__(/*! react-transform-catch-errors */ 29);\n\nvar _reactTransformCatchErrors4 = _interopRequireDefault(_reactTransformCatchErrors3);\n\nvar _react2 = __webpack_require__(/*! react */ 1);\n\nvar _react3 = _interopRequireDefault(_react2);\n\nvar _reactTransformHmr3 = __webpack_require__(/*! react-transform-hmr */ 30);\n\nvar _reactTransformHmr4 = _interopRequireDefault(_reactTransformHmr3);\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar _components = {\n  NoMatch: {\n    displayName: 'NoMatch'\n  }\n};\n\nvar _reactTransformHmr2 = (0, _reactTransformHmr4.default)({\n  filename: 'C:/Users/ColeChamberlain/cchamberlain/redux-webpack-boilerplate/src/app/routes/routes/NoMatch/components/NoMatch.js',\n  components: _components,\n  locals: [module],\n  imports: [_react3.default]\n});\n\nvar _reactTransformCatchErrors2 = (0, _reactTransformCatchErrors4.default)({\n  filename: 'C:/Users/ColeChamberlain/cchamberlain/redux-webpack-boilerplate/src/app/routes/routes/NoMatch/components/NoMatch.js',\n  components: _components,\n  locals: [],\n  imports: [_react3.default, _redboxReact3.default]\n});\n\nfunction _wrapComponent(id) {\n  return function (Component) {\n    return _reactTransformHmr2(_reactTransformCatchErrors2(Component, id), id);\n  };\n}\n\n/**\r\n * NoMatch Module\r\n * @module NoMatch\r\n */\n\nvar NoMatch = _wrapComponent('NoMatch')(function (_Component) {\n  _inherits(NoMatch, _Component);\n\n  function NoMatch() {\n    _classCallCheck(this, NoMatch);\n\n    return _possibleConstructorReturn(this, Object.getPrototypeOf(NoMatch).apply(this, arguments));\n  }\n\n  _createClass(NoMatch, [{\n    key: 'render',\n    value: function render() {\n      return _react3.default.createElement('div', null);\n    }\n  }]);\n\n  return NoMatch;\n}(_react2.Component));\n\nexports.default = NoMatch;\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(/*! ./../../../../../../~/webpack/buildin/module.js */ 25)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ../src/app/routes/routes/NoMatch/components/NoMatch.js\n ** module id = 421\n ** module chunks = 7\n **/\n//# sourceURL=webpack:///../src/app/routes/routes/NoMatch/components/NoMatch.js?");
 
 /***/ }
 
