@@ -1,12 +1,12 @@
 webpackJsonp([4,1],{
 
-/***/ 424:
+/***/ 427:
 /*!**********************************************************************!*\
   !*** ../src/app/routes/routes/Test/routes/Grids/components/Grids.js ***!
   \**********************************************************************/
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(/*! react */ 1);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _PrimaryGrid = __webpack_require__(/*! app/elements/grids/PrimaryGrid */ 205);\n\nvar _PrimaryGrid2 = _interopRequireDefault(_PrimaryGrid);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nexports.default = function (_ref) {\n  var children = _ref.children;\n  var params = _ref.params;\n  return _react2.default.createElement(\n    'div',\n    null,\n    _react2.default.createElement(\n      'h2',\n      null,\n      'Grid Tests'\n    ),\n    params.id ? _react2.default.createElement(\n      'span',\n      null,\n      params.id\n    ) : null,\n    _react2.default.createElement(_PrimaryGrid2.default, null),\n    children\n  );\n};\n\n/*****************\n ** WEBPACK FOOTER\n ** ../src/app/routes/routes/Test/routes/Grids/components/Grids.js\n ** module id = 424\n ** module chunks = 4\n **/\n//# sourceURL=webpack:///../src/app/routes/routes/Test/routes/Grids/components/Grids.js?");
+	eval("'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(/*! react */ 1);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _PrimaryGrid = __webpack_require__(/*! app/elements/grids/PrimaryGrid */ 204);\n\nvar _PrimaryGrid2 = _interopRequireDefault(_PrimaryGrid);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nexports.default = function (_ref) {\n  var children = _ref.children;\n  var params = _ref.params;\n  return _react2.default.createElement(\n    'div',\n    null,\n    _react2.default.createElement(\n      'h2',\n      null,\n      'Grid Tests'\n    ),\n    params.id ? _react2.default.createElement(\n      'span',\n      null,\n      params.id\n    ) : null,\n    _react2.default.createElement(_PrimaryGrid2.default, null),\n    children\n  );\n};\n\n/*****************\n ** WEBPACK FOOTER\n ** ../src/app/routes/routes/Test/routes/Grids/components/Grids.js\n ** module id = 427\n ** module chunks = 4\n **/\n//# sourceURL=webpack:///../src/app/routes/routes/Test/routes/Grids/components/Grids.js?");
 
 /***/ }
 
