@@ -1,12 +1,12 @@
 webpackJsonp([6,1],{
 
-/***/ 420:
+/***/ 423:
 /*!********************************************************!*\
   !*** ../src/app/routes/routes/Test/components/Test.js ***!
   \********************************************************/
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(/*! react */ 1);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactRouter = __webpack_require__(/*! react-router */ 175);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nexports.default = function (_ref) {\n  var children = _ref.children;\n  return _react2.default.createElement(\n    'div',\n    null,\n    _react2.default.createElement(\n      'ul',\n      null,\n      _react2.default.createElement(\n        'li',\n        null,\n        _react2.default.createElement(\n          _reactRouter.Link,\n          { to: '/test/forms' },\n          'Forms'\n        )\n      ),\n      _react2.default.createElement(\n        'li',\n        null,\n        _react2.default.createElement(\n          _reactRouter.Link,\n          { to: '/test/grids' },\n          'Grids'\n        )\n      )\n    ),\n    children\n  );\n};\n\n/*****************\n ** WEBPACK FOOTER\n ** ../src/app/routes/routes/Test/components/Test.js\n ** module id = 420\n ** module chunks = 6\n **/\n//# sourceURL=webpack:///../src/app/routes/routes/Test/components/Test.js?");
+	eval("'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(/*! react */ 1);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactRouter = __webpack_require__(/*! react-router */ 176);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nexports.default = function (_ref) {\n  var children = _ref.children;\n  return _react2.default.createElement(\n    'div',\n    null,\n    _react2.default.createElement(\n      'ul',\n      null,\n      _react2.default.createElement(\n        'li',\n        null,\n        _react2.default.createElement(\n          _reactRouter.Link,\n          { to: '/test/forms' },\n          'Forms'\n        )\n      ),\n      _react2.default.createElement(\n        'li',\n        null,\n        _react2.default.createElement(\n          _reactRouter.Link,\n          { to: '/test/grids' },\n          'Grids'\n        )\n      )\n    ),\n    children\n  );\n};\n\n/*****************\n ** WEBPACK FOOTER\n ** ../src/app/routes/routes/Test/components/Test.js\n ** module id = 423\n ** module chunks = 6\n **/\n//# sourceURL=webpack:///../src/app/routes/routes/Test/components/Test.js?");
 
 /***/ }
 
